refactor(iota-data-ingestion-core): remove CheckpointReaderV1
